Use Go1.12 bits.Mul64 if available
